Eleven: Don't use a sticky service
If the user removes Eleven from recents while the service is active
in background with nothing being played, the service is immediately
restarted and the notification re-created, even if there wasn't one.
Since the service doesn't resume music playback when restarted after
being killed because of OOM, there's really no point in using a sticky
service, so just switch to a non-sticky service.

Change-Id: I2a14176e60f535981a27af3536bd89b6d443d20e